refactor(multisig): split message-signing into read-only + writable files
Address review: put the read-only message-signing interface in its own file
(mirroring the wallet-account read-only/writable split) and rename it so ReadOnly
is a suffix matching the file name.

- New src/multisig/multisig-message-signing-read-only.js: IMultisigMessageSigningReadOnly
  (getMessageProposal/getMessageProposals) + owns the MultisigMessageProposal typedef.
- src/multisig/multisig-message-signing.js: IMultisigMessageSigning now extends
  IMultisigMessageSigningReadOnly and keeps only the writable methods + MultisigSignature.
- Updated the /multisig barrel and split the hand-maintained .d.ts to match.
